+++ src/pages/auth/JoinPage/index.tsx
@@ -8,9 +8,10 @@ import WideButton from '@/components/common/Button/WideButton';
 import { VaildType } from '@/types/auth/authDataType';
 import { isValidEmail, isValidPassword } from '@/util/valid';
 import usePostJoinQuery from '@/hooks/query/auth/usePostJoinQuery';
+import Success from '@/components/common/Toast/ToastSuccess';
+import useRedirectAfterToast from '@/hooks/useRedirectAfterToast';
 
 const JoinPage = () => {
-  const navigate = useNavigate();
   const [email, setEmail, handleChangeEmail] = useInput('');
   const [password, setPassword, handleChangePassword] = useInput('');
   const [confirmPassword, setConfirmPassword, handleChangeConfirmPassword] =
@@ -30,6 +31,7 @@ const JoinPage = () => {
     useState<VaildType>(validObj);
 
   const { joinIsMutate, joinIsSuccess } = usePostJoinQuery();
+  const redirect = useRedirectAfterToast(joinIsSuccess, '/auth/login');
 
   // 이메일 유효성 검사
   useEffect(() => {
@@ -126,18 +128,18 @@ const JoinPage = () => {
     joinIsMutate({ email, password, name: nickname, phone: phoneNumber });
   };
 
-  useEffect(() => {
-    let countRefId: NodeJS.Timeout;
-    if (joinIsSuccess) {
-      countRefId = setTimeout(() => {
-        navigate('/auth/login');
-      }, 1500);
-    }
+  // useEffect(() => {
+  //   let countRefId: NodeJS.Timeout;
+  //   if (joinIsSuccess) {
+  //     countRefId = setTimeout(() => {
+  //       navigate('/auth/login');
+  //     }, 1500);
+  //   }
 
-    return () => {
-      clearTimeout(countRefId);
-    };
-  }, [joinIsSuccess]);
+  //   return () => {
+  //     clearTimeout(countRefId);
+  //   };
+  // }, [joinIsSuccess]);
 
   return (
     <>
@@ -223,29 +225,8 @@ const JoinPage = () => {
         </div>
       </form>
 
-      <div
-        role="alert"
-        className={`alert alert-success fixed left-0 top-0 flex  ${
-          joinIsSuccess ? 'translate-y-0' : '-translate-y-full'
-        } transition `}
-      >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          className="stroke-current shrink-0 h-6 w-6"
-          fill="none"
-          viewBox="0 0 24 24"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth="2"
-            d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"
-          />
-        </svg>
-        <span>회원가입이 완료되었습니다!</span>
-      </div>
+      <Success status={joinIsSuccess} />
     </>
   );
 };
-
 export default JoinPage;
